fix: Modify register_blueprint(api_attractions_bp)
